Remove unneeded stuff
