--- __test__/data/course-1.json
@@ -1,7 +1,0 @@
-{
-    "type": "Course",
-    "title": "A course with no recipes",
-    "platforms": "all",
-    "description": "This is just a stub for testing",
-    "recipes": []
-}
