--- __test__/data/snake-1.2.3.json
@@ -1,38 +1,0 @@
-{
-    "type": "Recipe",
-    "title": "Installing Snake",
-    "tool": "snake-1.2.3",
-    "description": "A thing for testing. Totally not a python.",
-    "checkables": [
-        {
-            "type": "Checkable",
-            "steps": [
-                {
-                    "type": "Step",
-                    "title": "Download",
-                    "description": "download something from somewhere"
-                },
-                {
-                    "type": "Step",
-                    "title": "Install",
-                    "description": "run somethingsomething in a terminal"
-                }
-            ]
-        },
-        {
-            "type": "Checkable",
-            "steps": [
-                {
-                    "type": "Step",
-                    "title": "Configure",
-                    "description": "Edit the file whatever and change something"
-                },
-                {
-                    "type": "Step",
-                    "title": "Set path",
-                    "description": "do something about $PATH"
-                }
-            ]
-        }
-    ]
-}
