--- lib/parsers/Automation.js
@@ -1,30 +1,0 @@
-/*
- * Copyright (c) 2015 Nordic Semiconductor ASA
- *
- * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
- */
-
-import AbstractParser from './AbstractParser';
-import execute from './execute';
-
-export default class Automation extends AbstractParser {
-    constructor(json, id) {
-        super(json, 'Automation', id, [{ name: 'commands', optional: true }]);
-        this.commands =
-            this.commands instanceof Array ? this.commands : [this.commands];
-    }
-
-    run() {
-        return execute(
-            AbstractParser.replaceVariables(this.commands),
-            AbstractParser.getVariables()
-        );
-    }
-
-    serialize() {
-        return {
-            checkerType: this.checkerType,
-            command: this.command,
-        };
-    }
-}
