--- resources/data/arm_emb_gcc-0.1.3.json
@@ -1,35 +1,0 @@
-{
-    "type": "Recipe",
-    "title": "Install the toolchain",
-    "tool": "arm_emb_gcc-0.1.3",
-    "description": [
-        "Download and install version `9-2019-q4-major` of the GNU ARM Embedded toolchain to be able to cross-compile for ARM targets.",
-        "For more information, see [Installing the nRF Connect SDK manually](https://developer.nordicsemi.com/nRF_Connect_SDK/doc/latest/nrf/gs_installing.html#install-the-toolchain) in the nRF Connect SDK documentation."
-    ],
-    "checkables": [
-        {
-            "type": "Checkable",
-            "steps": [
-                {
-                    "type": "Step",
-                    "title": "Install cross compiler toolchain for ARM",
-                    "description": [
-                        "Download and install version `9-2019-q4-major` of the GNU ARM Embedded toolchain to be able to cross-compile for ARM targets.",
-                        "Extract the downloaded file to your preferred folder, for example, `/opt/gnuarmemb`.",
-                        "- [Linux x86 64-bit Tarball](https://developer.arm.com/-/media/Files/downloads/gnu-rm/9-2019q4/gcc-arm-none-eabi-9-2019-q4-major-x86_64-linux.tar.bz2)",
-                        "- [Linux AArch64 Tarball](https://developer.arm.com/-/media/Files/downloads/gnu-rm/9-2019q4/gcc-arm-none-eabi-9-2019-q4-major-aarch64-linux.tar.bz2)"
-                    ]
-                }
-            ],
-            "checkers": [
-                {
-                    "type": "Checker",
-                    "checkerType": "fileExists",
-                    "filenames": [
-                        "<GnuArmEmb_path:/opt/gnuarmemb>/bin/arm-none-eabi-gcc"
-                    ]
-                }
-            ]
-        }
-    ]
-}
