--- resources/data/nrf_connect_sdk-0.1.3.json
@@ -1,63 +1,0 @@
-{
-    "type": "Recipe",
-    "title": "Get the nRF Connect SDK code",
-    "tool": "nrf_connect_sdk-0.1.5",
-    "description": [
-        "The nRF Connect SDK consists of four GitHub repositories. To set up the nRF Connect SDK, you must clone these repositories and install the required Python modules.",
-        "If you intend to contribute directly to any of the repositories, you need to fork the repositories. For more information, see [Forking a repository of the nRF Connect SDK](https://developer.nordicsemi.com/nRF_Connect_SDK/doc/latest/nrf/dm_managing_code.html#forking-a-repository-of-the-ncs) in the nRF Connect SDK documentation."
-    ],
-    "checkables": [
-        {
-            "type": "Checkable",
-            "steps": [
-                {
-                    "type": "Step",
-                    "title": "Clone the repositories",
-                    "description": [
-                        "Initialize west and clone the nRF Connect SDK manifest repository `nrf`:",
-                        {
-                            "type": "commands",
-                            "description": [
-                                "cd <sourcecode_root>",
-                                "mkdir ncs",
-                                "cd ncs",
-                                "west init -m https://github.com/nrfconnect/sdk-nrf",
-                                "west update",
-                                "west zephyr-export"
-                            ]
-                        }
-                    ]
-                }
-            ],
-            "checkers": [
-                {
-                    "type": "Checker",
-                    "checkerType": "fileExists",
-                    "filenames": ["<sourcecode_root>/ncs/.west/config"]
-                }
-            ]
-        },
-        {
-            "type": "Checkable",
-            "steps": [
-                {
-                    "type": "Step",
-                    "title": "Check out the release",
-                    "description": [
-                        "Decide if you want to work with a tagged release of the nRF Connect SDK or with the latest state of development.",
-                        "_Note: The latest state is not necessarily tested. For a higher degree of quality assurance, check out a tagged release._",
-                        "To work with a tagged release, check the [Release Notes](https://developer.nordicsemi.com/nRF_Connect_SDK/doc/latest/nrf/release_notes.html#release-notes) for that release to find out the corresponding tag. Check out this tag in the `nrf` repository and use west to update the other repositories:",
-                        {
-                            "type": "commands",
-                            "description": [
-                                "cd <sourcecode_root>/ncs/nrf",
-                                "git checkout <ncs_tag>",
-                                "west update"
-                            ]
-                        }
-                    ]
-                }
-            ]
-        }
-    ]
-}
