--- resources/data/python_deps-0.1.3.json
@@ -1,31 +1,0 @@
-{
-    "type": "Recipe",
-    "title": "Install additional Python dependencies",
-    "tool": "python_deps-0.1.3",
-    "description": [
-        "The nRF Connect SDK Zephyr repository contains a list of required Python modules."
-    ],
-    "checkables": [
-        {
-            "type": "Checkable",
-            "steps": [
-                {
-                    "type": "Step",
-                    "title": "Install Python modules",
-                    "description": [
-                        "Install these modules by entering the following command:",
-                        {
-                            "type": "commands",
-                            "description": [
-                                "cd <sourcecode_root>/ncs",
-                                "pip3 install --user -r zephyr/scripts/requirements.txt",
-                                "pip3 install --user -r nrf/scripts/requirements.txt",
-                                "pip3 install --user -r bootloader/mcuboot/scripts/requirements.txt"
-                            ]
-                        }
-                    ]
-                }
-            ]
-        }
-    ]
-}
