--- resources/data/ses-0.1.1.json
@@ -1,29 +1,0 @@
-{
-    "type": "Recipe",
-    "title": "Install SEGGER Embedded Studio",
-    "tool": "ses-0.1.1",
-    "description": "The recommended way of building and programming an nRF Connect SDK sample is to use the **Nordic Edition** of the **SEGGER Embedded Studio (SES)** IDE. SEGGER Embedded Studio is free of charge for use with Nordic Semiconductor devices.",
-    "checkables": [
-        {
-            "type": "Checkable",
-            "steps": [
-                {
-                    "type": "Step",
-                    "platforms": { "platform": "linux" },
-                    "title": "Download",
-                    "description": "Download [SEGGER Embedded Studio](http://segger.com/downloads/embedded-studio/embeddedstudio_arm_nordic_linux_x64)."
-                }
-            ]
-        },
-        {
-            "type": "Checkable",
-            "steps": [
-                {
-                    "type": "Step",
-                    "title": "Extract",
-                    "description": "Extract SEGGER Embedded Studio to your preferred folder."
-                }
-            ]
-        }
-    ]
-}
