--- resources/data/ses_config-0.1.2.json
@@ -1,63 +1,0 @@
-{
-    "type": "Recipe",
-    "title": "Set up a project in SEGGER Embedded Studio",
-    "tool": "ses_config-0.1.3",
-    "description": "SEGGER Embedded Studio supports nRF Connect SDK projects. Note that you need the Nordic Edition of SEGGER Embedded Studio as described in the previous step.",
-    "checkables": [
-        {
-            "type": "Checkable",
-            "steps": [
-                {
-                    "type": "Step",
-                    "title": "Settings",
-                    "description": [
-                        "Navigate to the folder where you extracted SEGGER Embedded Studio. In the ``bin`` subfolder, run ``emStudio``.",
-                        "Click **Tools > Options** and select the **nRF Connect** tab.",
-                        "To load and build a project in SES, you must specify the following information:",
-                        "- **Zephyr Base** - the location of your cloned Zephyr repository (the full path to ``ncs/zephyr``)",
-                        "- **GNU ARM Embedded Toolchain Directory** - the location of your GNU ARM Embedded Toolchain installation (for example, ``c:\\gnuarmemb``)"
-                    ]
-                }
-            ]
-        },
-        {
-            "type": "Checkable",
-            "steps": [
-                {
-                    "type": "Step",
-                    "title": "Import project",
-                    "description": [
-                        "Select **File > Open nRF Connect SDK Project**.",
-                        "To import a project into SES, you must specify the following information:",
-                        "- **nRF Connect SDK Release** - Select the nRF Connect SDK version that you want to work with.",
-                        "- **nRF Connect Toolchain Version** - Select NONE.",
-                        "- **Projects** - Select the project that you want to work with.",
-                        "    Select any of the checkboxes to add samples and applications from the respective area to the drop-down list. To add custom projects to the drop-down list, click and select the folder that contains the projects that you want to add.",
-                        "- **Board Name** - Select the board that you want to work with.",
-                        "    Select any of the checkboxes to add the respective build targets to the drop-down list. To add custom build targets to the drop-down list, click and select the folder that contains the board definitions.",
-                        "    _Note that if you work with a TrustZone device, you usually build the firmware for the non-secure domain and must therefore select the non-secure build target, for example, **nrf9160dk_nrf9160ns** or **nrf5340pdk_nrf5340_cpuappns**._",
-                        "- **Build Directory** - Select the folder in which to run the build. The field is filled automatically based on the selected board name, but you can specify a different directory.",
-                        "- **Clean Build Directory** - Select this option to ensure that you are not building with an outdated build cache.",
-                        "- **Extended Settings** - Select this option to display a field where you can specify additional CMake options to be used for building.",
-                        "&nbsp;",
-                        "Click **OK** to open the project."
-                    ]
-                }
-            ]
-        },
-        {
-            "type": "Checkable",
-            "steps": [
-                {
-                    "type": "Step",
-                    "title": "Read user guides",
-                    "description": [
-                        "See the following user guides in the nRF Connect SDK documentation for essential information about developing applications for your device:",
-                        "- [Working with nRF9160](https://developer.nordicsemi.com/nRF_Connect_SDK/doc/latest/nrf/ug_nrf9160.html)",
-                        "- [Working with nRF5340](https://developer.nordicsemi.com/nRF_Connect_SDK/doc/latest/nrf/ug_nrf5340.html)"
-                    ]
-                }
-            ]
-        }
-    ]
-}
