--- resources/data/west-0.1.3.json
@@ -1,35 +1,0 @@
-{
-    "type": "Recipe",
-    "title": "Install west",
-    "tool": "west-0.1.3",
-    "description": [
-        "To manage the combination of repositories and versions, the nRF Connect SDK uses [Zephyr's west](https://developer.nordicsemi.com/nRF_Connect_SDK/doc/latest/zephyr/guides/west/index.html#west)."
-    ],
-    "checkables": [
-        {
-            "type": "Checkable",
-            "steps": [
-                {
-                    "type": "Step",
-                    "platforms": { "platform": "linux" },
-                    "title": "Install west",
-                    "description": [
-                        "Install west by entering the following command:",
-                        {
-                            "type": "commands",
-                            "description": ["pip3 install --user west"]
-                        }
-                    ]
-                }
-            ],
-            "checkers": [
-                {
-                    "type": "Checker",
-                    "checkerType": "command",
-                    "platforms": { "platform": "linux" },
-                    "filenames": ["west --version"]
-                }
-            ]
-        }
-    ]
-}
